Refactor upload to avatar controller
