+++ app/views/partials/detail.user.jade
@@ -1,5 +1,5 @@
 h2 {{user.name.fullName}}
-img(src="{{user.avatar}}")
+img(src="{{user.avatar.medium}}")
 h4 {{user.bio}}
 div.listing#selling
     table
